+++ src/model/model.ts
@@ -35,6 +35,7 @@
 export {Document, FeatureKinds, QueryOptions as DocumentQueryOptions, ScannedDocument} from './document';
 export * from './element';
 export {ElementReference, ScannedElementReference} from './element-reference';
+export * from './element-mixin';
 export * from './event';
 export * from './feature';
 export * from './import';