+++ src/test/static/polymer2/test-mixin-1.js
@@ -1,0 +1,17 @@
+/**
+ * @polymerMixin
+ */
+function TestMixin(superclass) {
+  return class extends superclass {
+    static get config() {
+      return {
+        properties: {
+          foo: {
+            notify: true,
+            type: String,
+          }
+        },
+      };
+    }
+  }
+}